requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chaincomposer==0.1.8
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docx2txt==0.8
emoji==2.14.1
eval-type-backport==0.2.2
fastapi==0.115.10
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-anthropic==0.3.0
langchain-community==0.3.18
langchain-core==0.3.40
langchain-google-genai==2.0.5
langchain-openai==0.2.9
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.1.147
lxml==5.3.1
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.58.1
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pymupdf==1.25.3
pyparsing==3.2.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==1.0.2
pyyaml==6.0.2
rapidfuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.38
starlette==0.46.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
unstructured==0.16.23
unstructured-client==0.30.6
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
wrapt==1.17.2
xlsxwriter==3.2.2
yarl==1.18.3
